requirements.txt
Flask==1.1.2
pymongo[tls,srv]==3.11.2
python-dotenv==0.15.0
requests==2.25.1